feat(FX-3295): Added no results state when Algolia query comes back empty (#5440)
* FX-3295 Added state when no results

* FX-3295 Separated SearchResults component
