+++ src/lib/Scenes/Search2/Search2.tsx
@@ -1,29 +1,24 @@
 import { captureMessage } from "@sentry/react-native"
 import { Search2_system } from "__generated__/Search2_system.graphql"
 import { Search2Query } from "__generated__/Search2Query.graphql"
-import { AboveTheFoldFlatList } from "lib/Components/AboveTheFoldFlatList"
 import { ArtsyKeyboardAvoidingView } from "lib/Components/ArtsyKeyboardAvoidingView"
-import OpaqueImageView from "lib/Components/OpaqueImageView/OpaqueImageView"
 import { SearchInput as SearchBox } from "lib/Components/SearchInput"
-import { navigate, navigateToPartner } from "lib/navigation/navigate"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
 import { isPad } from "lib/utils/hardware"
 import { ProvidePlaceholderContext } from "lib/utils/placeholders"
 import { Schema } from "lib/utils/track"
 import { useAlgoliaClient } from "lib/utils/useAlgoliaClient"
-import { searchInsights, useSearchInsightsConfig } from "lib/utils/useSearchInsightsConfig"
-import { Flex, Pill, Spacer, Spinner, Text, Touchable, useSpace } from "palette"
-import React, { useEffect, useMemo, useRef, useState } from "react"
-import { InfiniteHitsProvided, StateResultsProvided } from "react-instantsearch-core"
+import { useSearchInsightsConfig } from "lib/utils/useSearchInsightsConfig"
+import { Flex, Pill, Spacer } from "palette"
+import React, { useMemo, useState } from "react"
 import {
   Configure,
-  connectHighlight,
   connectInfiniteHits,
   connectSearchBox,
   connectStateResults,
   InstantSearch,
 } from "react-instantsearch-native"
-import { FlatList, Platform, ScrollView } from "react-native"
+import { Platform, ScrollView } from "react-native"
 import { createRefetchContainer, graphql, QueryRenderer, RelayRefetchProp } from "react-relay"
 import { useTracking } from "react-tracking"
 import styled from "styled-components"
@@ -34,7 +29,7 @@ import { SearchContext, useSearchProviderValues } from "../Search/SearchContext"
 import { SearchPlaceholder } from "./components/SearchPlaceholder"
 import { RefetchWhenApiKeyExpiredContainer } from "./containers/RefetchWhenApiKeyExpired"
 import { SearchArtworksGridQueryRenderer } from "./SearchArtworksGrid"
-import { AlgoliaSearchResult } from "./types"
+import { SearchResults } from "./SearchResults"
 
 interface SearchInputProps {
   placeholder: string
@@ -73,104 +68,6 @@ const SearchInput: React.FC<SearchInputProps> = ({ currentRefinement, placeholde
   )
 }
 
-const Highlight = connectHighlight(({ highlight, attribute, hit, highlightProperty = "_highlightResult" }) => {
-  const parsedHit = highlight({ attribute, hit, highlightProperty })
-
-  return (
-    <Text>
-      {parsedHit.map(({ isHighlighted, value }, index) =>
-        isHighlighted ? (
-          <Text key={index} color="blue100" fontWeight="600" padding={0} margin={0}>
-            {value}
-          </Text>
-        ) : (
-          <Text key={index}>{value}</Text>
-        )
-      )}
-    </Text>
-  )
-})
-
-interface SearchResultsProps
-  extends StateResultsProvided<AlgoliaSearchResult>,
-    InfiniteHitsProvided<AlgoliaSearchResult> {
-  indexName: string
-}
-
-const SearchResults: React.FC<SearchResultsProps> = ({
-  hits,
-  hasMore,
-  searching,
-  isSearchStalled,
-  searchState,
-  indexName,
-  refineNext,
-}) => {
-  const flatListRef = useRef<FlatList<AlgoliaSearchResult>>(null)
-  const loading = searching || isSearchStalled
-  const space = useSpace()
-
-  useEffect(() => {
-    flatListRef.current?.scrollToOffset({ offset: 1, animated: true })
-  }, [searchState.query, indexName])
-
-  const onPress = (item: AlgoliaSearchResult): void => {
-    // TODO: I'm not sure why we need to use this `navigateToPartner` function but without it the header overlaps
-    // with the back button
-    if (item.href.startsWith("/partner/")) {
-      navigateToPartner(item.slug)
-    } else {
-      navigate(item.href)
-    }
-
-    searchInsights("clickedObjectIDsAfterSearch", {
-      index: indexName,
-      eventName: "Search item clicked",
-      positions: [item.__position],
-      queryID: item.__queryID,
-      objectIDs: [item.objectID],
-    })
-  }
-
-  const loadMore = () => {
-    if (hasMore && !loading) {
-      refineNext()
-    }
-  }
-
-  return (
-    <AboveTheFoldFlatList<AlgoliaSearchResult>
-      listRef={flatListRef}
-      initialNumToRender={isPad() ? 24 : 12}
-      contentContainerStyle={{ paddingVertical: space(1) }}
-      data={hits}
-      keyExtractor={(item) => item.objectID}
-      renderItem={({ item }) => (
-        <Touchable onPress={() => onPress(item)}>
-          <Flex py={space(1)} px={space(2)} flexDirection="row" alignItems="center">
-            <OpaqueImageView
-              imageURL={item.image_url}
-              style={{ width: 40, height: 40, borderRadius: 20, overflow: "hidden" }}
-            />
-            <Spacer ml={1} />
-            <Flex>
-              <Highlight attribute="name" hit={item} />
-            </Flex>
-          </Flex>
-        </Touchable>
-      )}
-      onEndReached={loadMore}
-      ListFooterComponent={
-        <Flex alignItems="center" my={2}>
-          {loading ? <Spinner /> : null}
-        </Flex>
-      }
-      keyboardDismissMode="on-drag"
-      keyboardShouldPersistTaps="handled"
-    />
-  )
-}
-
 const SearchInputContainer = connectSearchBox(SearchInput)
 const SearchResultsContainerWithState = connectStateResults(SearchResults)
 const SearchResultsContainer = connectInfiniteHits(SearchResultsContainerWithState)
@@ -180,7 +77,12 @@ interface SearchState {
   page?: number
 }
 
-const pills = [{ name: "ARTWORK", displayName: "Artworks" }]
+interface PillType {
+  name: string
+  displayName: string
+}
+
+const pills: PillType[] = [{ name: "ARTWORK", displayName: "Artworks" }]
 
 interface Search2Props {
   relay: RelayRefetchProp
@@ -193,10 +95,11 @@ export const Search2: React.FC<Search2Props> = (props) => {
   const [selectedAlgoliaIndex, setSelectedAlgoliaIndex] = useState("")
   const [elasticSearchEntity, setElasticSearchEntity] = useState("")
   const searchProviderValues = useSearchProviderValues(searchState?.query ?? "")
+  const [activePillDisplayName, setActivePillDisplayName] = useState("")
   const { searchClient } = useAlgoliaClient(system?.algolia?.appID!, system?.algolia?.apiKey!)
   const searchInsightsConfigured = useSearchInsightsConfig(system?.algolia?.appID, system?.algolia?.apiKey)
 
-  const pillsArray = useMemo(() => {
+  const pillsArray = useMemo<PillType[]>(() => {
     const indices = system?.algolia?.indices
 
     if (Array.isArray(indices) && indices.length > 0) {
@@ -214,9 +117,9 @@ export const Search2: React.FC<Search2Props> = (props) => {
     )
   }
 
-  const renderResults = () => {
+  const renderResults = (categoryDisplayName: string) => {
     if (!!selectedAlgoliaIndex) {
-      return <SearchResultsContainer indexName={selectedAlgoliaIndex} />
+      return <SearchResultsContainer indexName={selectedAlgoliaIndex} categoryDisplayName={categoryDisplayName} />
     }
     if (!!elasticSearchEntity) {
       return <SearchArtworksGridQueryRenderer keyword={searchState.query!} />
@@ -226,7 +129,8 @@ export const Search2: React.FC<Search2Props> = (props) => {
 
   const shouldStartQuering = !!searchState?.query?.length && searchState?.query.length >= 2
 
-  const handlePillPress = (name: string) => {
+  const handlePillPress = ({ name, displayName }: PillType) => {
+    setActivePillDisplayName(displayName)
     setSearchState((prevState) => ({ ...prevState, page: 1 }))
     if (name === "ARTWORK") {
       setSelectedAlgoliaIndex("")
@@ -255,20 +159,23 @@ export const Search2: React.FC<Search2Props> = (props) => {
             <>
               <Flex p={2} pb={1}>
                 <ScrollView horizontal showsHorizontalScrollIndicator={false}>
-                  {pillsArray.map(({ name, displayName }) => (
-                    <Pill
-                      mr={1}
-                      key={name}
-                      rounded
-                      selected={selectedAlgoliaIndex === name || elasticSearchEntity === name}
-                      onPress={() => handlePillPress(name)}
-                    >
-                      {displayName}
-                    </Pill>
-                  ))}
+                  {pillsArray.map((pill) => {
+                    const { name, displayName } = pill
+                    return (
+                      <Pill
+                        mr={1}
+                        key={name}
+                        rounded
+                        selected={selectedAlgoliaIndex === name || elasticSearchEntity === name}
+                        onPress={() => handlePillPress(pill)}
+                      >
+                        {displayName}
+                      </Pill>
+                    )
+                  })}
                 </ScrollView>
               </Flex>
-              {renderResults()}
+              {renderResults(activePillDisplayName)}
             </>
           ) : (
             <Scrollable>
